+++ Software/Java/RabbitMQClient/bin/start-rabbitmq.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
-
+# Starts a RabbitMQ Server in a Docker container which can be used for testing
+# The settings used by this server should be compatible with the default
+# settings in mps.yaml and example.yaml.
 set -x
 
 docker run -d --rm \
-    --hostname l-shaped-rabbit \
-    --name math-savvy-rabbit \
+    --name rabbitmq-for-matlab \
     -p 5672:5672 \
     -p 15672:15672 \
-    -e RABBITMQ_DEFAULT_VHOST=rabbit \
     rabbitmq:3-management