+++ Software/MATLAB/app/system/+rabbitmq/@object/object.m
@@ -1,0 +1,16 @@
+classdef object < dynamicprops
+    % OBJECT Root Class for all RabbitMQ wrapper objects
+    
+    % Copyright 2020-2022 The MathWorks, Inc
+    
+    properties (Hidden)
+        Handle;
+    end
+    
+    methods
+        %% Constructor
+        function obj = object(~, varargin)
+        end
+    end
+    
+end %class