+++ RELEASENOTES.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## Release 0.2.0 (23rd March 2022)
+* Fixed library dependency conflicts and updated library dependencies
+* Renamed `MessageReceiver` to `MessageBroker` to underline the fact that the
+  application does not only receive messages but then also actually forwards
+  them to MATLAB Production Server
+* Added a MATLAB interface for interacting with RabbitMQ from MATLAB directly
+* Documentation improvements
+* Code clean-up
+
 ## Release 0.1.4 (13th Oct 2020)
 * Bumped junit version
 
@@ -17,4 +26,4 @@
 ## Release 0.1.0 (Dec 2019)
 * Initial release supporting RabbitMQ
 
-[//]: #  (Copyright 2019-2020 The MathWorks, Inc.)
+[//]: #  (Copyright 2019-2022 The MathWorks, Inc.)